refactored a few things
