+++ src/containers/login/index.tsx
@@ -8,6 +8,7 @@ import { AsyncRequestKinds } from '../../utils/asyncRequest';
 import { C4CState } from '../../store';
 import { UserAuthenticationReducerState } from '../../auth/ducks/types';
 import { ContentContainer } from '../../components';
+import {Routes} from '../../App';
 
 const { Title, Paragraph } = Typography;
 
@@ -43,14 +44,14 @@ const Login: React.FC<LoginProps> = ({ tokens }) => {
           </Form.Item>
           <Paragraph>
             Need an account? Sign up{' '}
-            <Link to="/signup" component={Typography.Link}>
+            <Link to={Routes.SIGNUP} component={Typography.Link}>
               here
             </Link>
             !
           </Paragraph>
           <Paragraph>
             Forgot your password? Click{' '}
-            <Link to="/forgot-password" component={Typography.Link}>
+            <Link to={Routes.FORGOT_PASSWORD_REQUEST} component={Typography.Link}>
               here
             </Link>{' '}
             to reset it.
